separate base word counter functionality and realtime word counter classses
